requirements.txt
faiss-cpu
numpy
tqdm
python-dotenv
sentence-transformers
openai